Initial
This project consists of 3 header files and one main file but if we talk about classes that are eight classes piece class is a base class, and all other classes are derived from piece class. Actual working is present in the Tetris class that runs the game and all other information or you can say that Tetris class is the main class.